refactor(utils): split utils and tests
