+++ .gitignore
@@ -23,3 +23,5 @@ test_dir
 .mypy_cache
 .pytest_cache
 codealike.json
+
+junit
