+++ bgmi/front/index.py
@@ -9,7 +9,7 @@ from bgmi.config import FRONT_STATIC_PATH, SAVE_PATH
 from bgmi.front.base import COVER_URL, BaseHandler
 from bgmi.lib.db_models import Bangumi, Followed
 from bgmi.logger import logger
-from bgmi.utils import normalize_path
+from bgmi.pure_utils import normalize_path
 
 
 def get_player(bangumi_name):
