+++ bgmi/models/followed.py
@@ -4,9 +4,9 @@ from pydantic import BaseModel
 
 from bgmi import config
 from bgmi.lib import db_models
-from bgmi.lib.db_models._tables import split_str_to_list
 from bgmi.models import Episode
 from bgmi.models.status import FollowedStatus
+from bgmi.pure_utils import split_str_to_list
 from bgmi.utils.followed_filter import apply_regex
