+++ bgmi/utils/__init__.py
@@ -1,9 +1,8 @@
 from ._decorator import COLOR_END, GREEN, GREY, RED, WHITE, YELLOW
 from .episode_parser import chinese_to_arabic, parse_episode
 from .utils import (
-    check_update, convert_cover_url_to_path, download_cover, exec_command, full_to_half,
-    get_terminal_col, get_web_admin, normalize_path, parallel, print_error, print_info,
-    print_success, print_warning, render_template, test_connection
+    check_update, convert_cover_url_to_path, download_cover, get_terminal_col, get_web_admin,
+    print_error, print_info, print_success, print_warning
 )
 
 __all__ = [
@@ -13,21 +12,15 @@ __all__ = [
     'RED',
     'WHITE',
     'YELLOW',
-    'parallel',
     'chinese_to_arabic',
     'parse_episode',
     'check_update',
     'convert_cover_url_to_path',
     'download_cover',
-    'exec_command',
-    'full_to_half',
     'get_terminal_col',
     'get_web_admin',
-    'normalize_path',
     'print_error',
     'print_info',
     'print_success',
     'print_warning',
-    'render_template',
-    'test_connection',
 ]
